add not found partial to be used on business case and system intake
