+++ src/views/SystemIntake/index.tsx
@@ -16,6 +16,7 @@ import {
   storeSystemIntake
 } from 'types/routines';
 import { SystemIntakeForm } from 'types/systemIntake';
+import { NotFoundPartial } from 'views/NotFound';
 
 import ContactDetails from './ContactDetails';
 import RequestDetails from './RequestDetails';
@@ -84,14 +85,6 @@ export const SystemIntake = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  useEffect(() => {
-    const pageSlugs = ['contact-details', 'request-details', 'review'];
-    if (!pageSlugs.includes(formPage)) {
-      history.replace(`/system/${systemId}/contact-details`);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [systemId, formPage]);
-
   return (
     <div className="system-intake margin-bottom-5">
       <Header />
@@ -150,6 +143,7 @@ export const SystemIntake = () => {
               path="/system/:systemId/review"
               render={() => <Review systemIntake={systemIntake} />}
             />
+            <SecureRoute path="*" render={() => <NotFoundPartial />} />
           </>
         )}
       </MainContent>
